+++ docs/src/examples/SidebarStructure/GroupsExample.js
@@ -7,25 +7,26 @@ import ShardDocs from "@fa-repo/shard-docs";
 
 const SidebarStructureGroupsExample = props => (
   <ShardDocs
+    title="Groups example"
     basePath="/examples/sidebar-structure-groups"
     tree={[
-      { page: "Page A", composition: [<h1>Page A</h1>] },
-      { page: "Page B", composition: [<h1>Page A</h1>] },
-      { page: "Page C", composition: [<h1>Page A</h1>] },
+      { page: "Camels", composition: [<h1>Camels</h1>] },
+      { page: "Hippos", composition: [<h1>Hippos</h1>] },
+      { page: "Baboons", composition: [<h1>Baboons</h1>] },
       {
-        group: "Group A",
+        group: "Wine",
         pages: [
-          { page: "Page D", composition: [<h1>Page D</h1>] },
-          { page: "Page E", composition: [<h1>Page E</h1>] },
-          { page: "Page F", composition: [<h1>Page F</h1>] }
+          { page: "Red Bordeaux", composition: [<h1>Red Bordeaux</h1>] },
+          { page: "Champagne", composition: [<h1>Champagne</h1>] },
+          { page: "Rioja", composition: [<h1>Rioja</h1>] }
         ]
       },
       {
         group: null,
         pages: [
-          { page: "Page G", composition: [<h1>Page G</h1>] },
-          { page: "Page H", composition: [<h1>Page H</h1>] },
-          { page: "Page I", composition: [<h1>Page I</h1>] }
+          { page: "Great Western Rail", composition: [<h1>Great Western Rail</h1>] },
+          { page: "East Midlands Trains", composition: [<h1>East Midlands Trains</h1>] },
+          { page: "Eurostar", composition: [<h1>Eurostar</h1>] }
         ]
       }
     ]}