+++ docs/src/examples/SidebarStructure/SidebarStructure.js
@@ -17,10 +17,11 @@ export default [
   <CodeExampleShard
     sourceCode={`
 <ShardDocs
+  title="Basic example"
   tree={[
-    { page: "Page A", composition: [<h1>Page A</h1>] },
-    { page: "Page B", composition: [<h1>Page B</h1>] },
-    { page: "Page C", composition: [<h1>Page C</h1>] },
+    { page: "Mangos", composition: [<h1>Mangos</h1>] },
+    { page: "Plums", composition: [<h1>Plums</h1>] },
+    { page: "Avacadoes", composition: [<h1>Avacadoes</h1>] },
   ]}
 />`}
   >
@@ -30,22 +31,27 @@ export default [
   <CodeExampleShard
     sourceCode={`
 <ShardDocs
+  title="Groups example"
   tree={[
-    { page: "Page A", composition: [<h1>Page A</h1>] },
+    { page: "Camels", composition: [<h1>Camels</h1>] },
+    { page: "Hippos", composition: [<h1>Hippos</h1>] },
+    { page: "Baboons", composition: [<h1>Baboons</h1>] },
     {
-      group: "Group A",
+      group: "Wine",
       pages: [
-        { page: "Page B", composition: [<h1>Page B</h1>] },
-        { page: "Page C", composition: [<h1>Page C</h1>] }
+        { page: "Red Bordeaux", composition: [<h1>Red Bordeaux</h1>] },
+        { page: "Champagne", composition: [<h1>Champagne</h1>] },
+        { page: "Rioja", composition: [<h1>Rioja</h1>] }
       ]
     },
     {
-      group: null, /* discrete group */
+      group: null,
       pages: [
-        { page: "Page D", composition: [<h1>Page D</h1>] },
-        { page: "Page E", composition: [<h1>Page E</h1>] }
+        { page: "Great Western Rail", composition: [<h1>Great Western Rail</h1>] },
+        { page: "East Midlands Trains", composition: [<h1>East Midlands Trains</h1>] },
+        { page: "Eurostar", composition: [<h1>Eurostar</h1>] }
       ]
-    },
+    }
   ]}
 />`}
   >
@@ -56,35 +62,30 @@ export default [
   <CodeExampleShard
     sourceCode={`
 <ShardDocs
+  title="Indentation example"
   tree={[
     {
-      group: "Group A",
+      group: "Europe bucket list:",
       pages: [
-        { page: "Page B", composition: [<h1>Page B</h1>] },
-        { page: "Page C", composition: [<h1>Page C</h1>] },
         {
-          group: "Group B",
+          group: null,
           pages: [
-            { page: "Page D", composition: [<h1>Page D</h1>] },
-            { page: "Page E", composition: [<h1>Page E</h1>] }
+            {
+              page: "- France: Scale Mont Blanc",
+              composition: [<h1>Scale Mont Blanc</h1>]
+            },
+            {
+              page: "- Italy: Dine at a Tuscan restaurent",
+              composition: [<h1>Dine at a Tuscan restaurent</h1>]
+            },
+            {
+              page: "- Spain: Walk the Camino de Santiago",
+              composition: [<h1>Dine at a rustic Tuscan restaurent</h1>]
+            }
           ]
-        },
+        }
       ]
-    },
-    {
-      group: null, /* discrete group */
-      pages: [
-        { page: "Page G", composition: [<h1>Page F</h1>] },
-        { page: "Page G", composition: [<h1>Page G</h1>] },
-        {
-          group: null, /* discrete group */
-          pages: [
-            { page: "Page H", composition: [<h1>Page H</h1>] },
-            { page: "Page I", composition: [<h1>Page I</h1>] }
-          ]
-        },
-      ]
-    },
+    }
   ]}
 />`}
   >