+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Pages not showing up in discrete groups
+### Changed
+- Combine all top level adjacent pages into discrete groups by default.
 
 ## [4.0.0] - 2019-08-21
 ### Added